fix($TestInit): 초기화 메서드 버그 수정
기존의 @PostConstruct로 동작하던 초기화 메서드로 인해
스프링 빈이 제대로 생성되지 않는 문제가 발생함.
(스프링 빈의
의존관계 주입 순서와 @PostConstruct의 동작 시점 관련 문제)

이러한 문제를 @EventListener를 이용하여
해결함.
(application 로딩을 완전히 완료한 이후에 초기화 진행)
