+++ src/main/java/com/whatachad/app/init/OrderedPostConstruct.java
@@ -1,0 +1,5 @@
+package com.whatachad.app.init;
+
+public interface OrderedPostConstruct {
+    void init();
+}
